refactor:レビュー修正・Modelsディレクトリ作成してそれぞれファイル記載部名前変更。
