+++ src/config/auth.php
@@ -72,7 +72,7 @@ return [
     'providers' => [
         'admins' => [
             'driver' => 'eloquent',
-            'model' => App\AdminUser::class,
+            'model' => App\Models\AdminUser::class,
         ],
 
         'users' => [
